Module alloc_data::prelude

Crate’s prelude.

Re-exports

pub use crate::parser;

pub use crate::parser::Parseable;

pub use crate::err;

pub use crate::mem;

pub use crate::mem::str::Str;

pub use crate::Alloc;

pub use crate::AllocKind;

pub use crate::Builder;

pub use crate::CLoc;

pub use crate::Diff;

pub use crate::Init;

pub use crate::Loc;

pub use crate::Span;

Modules

err

Errors for memthol, handled by error_chain.

error_chain

Re-exports from error_chain.

fmt

Utilities for formatting and printing Strings.

log

Log macros re-exports.

ops

Overloadable operators.

serde

Serde trait re-exports.

sync

Useful synchronization primitives.

time

Structures representing time in different ways.

time_stats

Stopwatch, for time statistics.

uid

UID types for charts, filters and subfilters.

Macros

bail

Exits a function early with an error

lazy_static

Structs

Arc

A thread-safe reference-counting pointer. ‘Arc’ stands for ‘Atomically Reference Counted’.

BTMap

An ordered map based on a B-Tree.

BTSet

An ordered set based on a B-Tree.

BigUint

A big unsigned integer type.

HMap

A hash map implemented with quadratic probing and SIMD lookup.

HSet

A hash set implemented as a HashMap where the value is ().

Labels

Stores a UID, cannot be constructed outside of the module it’s declared in.

Range

A range, inclusive on both ends.

SampleRate

Represents a sample rate.

Trace

Stores a UID, cannot be constructed outside of the module it’s declared in.

Enums

Either

The enum Either with variants Left and Right is a general purpose sum type with two cases.

Inhabited

Inhabited type.

Traits

Borrow

A trait for borrowing data.

DurationExt

Adds functionalities to the Duration type.

FromStr

Parse a value from a string

ResExt

Additional methods for Result, for easy interaction with this crate.

TryFrom

Simple and safe type conversions that may fail in a controlled way under some circumstances. It is the reciprocal of TryInto.

TryInto

An attempted conversion that consumes self, which may or may not be expensive.

Functions

convert

Used to convert between integer representations.

destroy

Destroys what it’s given.

identity

Returns what it’s given.

Type Definitions

Res

Convenient wrapper around std::Result.

SVec8

Alias type for SmallVec of max stack-size 8.

SVec16

Alias type for SmallVec of max stack-size 16.

SVec32

Alias type for SmallVec of max stack-size 32.

SVec64

Alias type for SmallVec of max stack-size 64.

Derive Macros

Deserialize

Serialize