Struct alloc_data::prelude::time::SinceStart

pub struct SinceStart { /* private fields */ }

Wrapper around a duration.

This type represents a point in time relative to the start time of the run of the program being profiled, which is a Date (an absolute point in time).

Implementations

impl SinceStart

pub fn zero() -> SinceStart

A duration of 0 nanoseconds.

pub fn one_sec() -> SinceStart

A duration of 1 second.

pub fn from_secs(secs: u64) -> SinceStart

Constructor from an amount of seconds.

pub fn from_nano_timestamp(secs: u64, nanos: u32) -> SinceStart

Constructor from a timestamp in nanos seconds.

pub fn is_zero(&self) -> bool

True if the duration is zero.

pub fn to_lifetime(self) -> Lifetime

Turns itself in a lifetime.